auth-service/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Redis and caching
redis[hiredis]==5.0.1
aioredis==2.0.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
cryptography==41.0.7
python-multipart==0.0.6

# OAuth providers
authlib==1.2.1
httpx==0.25.2

# Email
aiosmtplib==3.0.1
jinja2==3.1.2

# Background tasks
celery[redis]==5.3.4
flower==2.0.1

# Configuration and environment
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Validation and serialization
pydantic[email]==2.5.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# CORS
python-cors==1.0.1

# Rate limiting
slowapi==0.1.9

# Template engine
jinja2==3.1.2

# File handling
python-magic==0.4.27

# Timezone handling
pytz==2023.3

# JSON Web Tokens
pyjwt==2.8.0

# Password strength validation
zxcvbn==4.4.24